generate metadata with git short commits (#67)
use a new makefile goal to generate an extra
yaml file in the `deploy/nstemplatetiers` directory,
which contains an entry for each template manifest,
along with the latest commit (short) hash.

This file is then included in the generated
`pkg/templates/templates_content.go` file
so the metadata can be retrieved later

this generated `deploy/nstemplatetiers/metadata.yaml`
is excluded from the git repository

also, uses github.com/src-d/go-git to verify the revisions/commit hashes. 

fixes CRT-262
